#ESW-414 : Shivani | Chauhan:  Update the template to implement the changes discussed
# Following changes were made
    1. organization should be removed. should be fixed to com.github.tmtsoftware.<project-name>
    2. the package should be <subsystem> instead of <organization>.<subsystem>
    3. the names of the modules(hcd, assembly, deploy) should include subsystem
