--- src/main/g8/$assemblyComponentName__space,norm$/src/test/java/$organization__lower,packaged$/$subsystem__lower$/$assemblyComponentName__space,norm,word$/J$assemblyComponentName__space,Camel$Test.java
@@ -1,41 +1,0 @@
-package $organization$.$subsystem;format="lower"$.$assemblyComponentName;format="space,norm,word"$;
-
-import csw.location.api.javadsl.ILocationService;
-import csw.location.api.javadsl.JComponentType;
-import csw.location.api.models.AkkaLocation;
-import csw.location.api.models.ComponentId;
-import csw.location.api.models.Connection;
-import csw.prefix.models.Prefix;
-import csw.testkit.javadsl.FrameworkTestKitJunitResource;
-import csw.testkit.javadsl.JCSWService;
-import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.Test;
-import org.scalatestplus.junit.JUnitSuite;
-
-import java.time.Duration;
-import java.util.Arrays;
-import java.util.concurrent.ExecutionException;
-
-public class J$assemblyComponentName;format="space,Camel"$Test extends JUnitSuite {
-
-    @ClassRule
-    public static final FrameworkTestKitJunitResource testKit =
-        new FrameworkTestKitJunitResource(Arrays.asList(JCSWService.AlarmServer, JCSWService.EventServer));
-
-    @BeforeClass
-    public static void setup() {
-        // uncomment if you want one Assembly run for all tests
-        testKit.spawnStandalone(com.typesafe.config.ConfigFactory.load("J$assemblyComponentName;format="space,Camel"$Standalone.conf"));
-    }
-
-    @Test
-    public void testAssemblyShouldBeLocatableUsingLocationService() throws ExecutionException, InterruptedException {
-        Connection.AkkaConnection connection = new Connection.AkkaConnection(new ComponentId(Prefix.apply("$subsystem$.$assemblyComponentName;format="package"$"), JComponentType.Assembly));
-        ILocationService locationService = testKit.jLocationService();
-        AkkaLocation location = locationService.resolve(connection, Duration.ofSeconds(10)).get().get();
-
-        Assert.assertEquals(location.connection(), connection);
-    }
-}
