--- src/main/g8/$assemblyComponentName__space,norm$/src/test/resources/$assemblyComponentName__space,Camel$Standalone.conf
@@ -1,4 +1,0 @@
-prefix = "$subsystem$.$assemblyComponentName;format="package"$"
-componentType = assembly
-behaviorFactoryClassName = "$organization$.$subsystem;format="lower"$.$assemblyComponentName;format="space,norm,word"$.$assemblyComponentName;format="space,Camel"$BehaviorFactory"
-locationServiceUsage = RegisterOnly
