--- src/main/g8/$hcdComponentName__space,norm$/src/test/scala/$organization__lower,packaged$/$subsystem__lower$/$hcdComponentName__space,norm$/$hcdComponentName__space,Camel$Test.scala
@@ -1,29 +1,0 @@
-package $organization$.$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$
-
-import csw.location.api.models.Connection.AkkaConnection
-import csw.location.api.models.{ComponentId, ComponentType}
-import csw.prefix.models.Prefix
-import csw.testkit.scaladsl.CSWService.{AlarmServer, EventServer}
-import csw.testkit.scaladsl.ScalaTestFrameworkTestKit
-import org.scalatest.funsuite.AnyFunSuiteLike
-
-import scala.concurrent.Await
-import scala.concurrent.duration._
-
-class $hcdComponentName;format="space,Camel"$Test extends ScalaTestFrameworkTestKit(AlarmServer, EventServer) with AnyFunSuiteLike {
-
-  import frameworkTestKit.frameworkWiring._
-
-  override def beforeAll(): Unit = {
-    super.beforeAll()
-    // uncomment if you want one HCD run for all tests
-    spawnStandalone(com.typesafe.config.ConfigFactory.load("$hcdComponentName;format="space,Camel"$Standalone.conf"))
-  }
-
-  test("HCD should be locatable using Location Service") {
-    val connection = AkkaConnection(ComponentId(Prefix("$subsystem$.$hcdComponentName;format="package"$"), ComponentType.HCD))
-    val akkaLocation = Await.result(locationService.resolve(connection, 10.seconds), 10.seconds).get
-
-    akkaLocation.connection shouldBe connection
-  }
-}
