--- src/main/g8/$name__space,norm$-deploy/README.md
@@ -1,39 +1,0 @@
-# $name;format="space,norm"$-deploy
-
-This module contains apps and configuration files for host deployment using 
-HostConfig (https://tmtsoftware.github.io/csw/apps/hostconfig.html) and 
-ContainerCmd (https://tmtsoftware.github.io/csw/framework/deploying-components.html).
-
-An important part of making this work is ensuring the host config app ($name;format="space,Camel"$HostConfigApp) is built
-with all of the necessary dependencies of the components it may run.  This is done by adding settings to the
-built.sbt file:
-
-```
-lazy val `$name;format="space,norm"$-deploy` = project
-  .dependsOn(
-    `$assemblyComponentName;format="space,norm"$`,
-    `$hcdComponentName;format="space,norm"$``
-  )
-  .enablePlugins(JavaAppPackaging)
-  .settings(
-    libraryDependencies ++= Dependencies.$name;format="space,Camel"$Deploy
-  )
-```
-
-and in Libs.scala:
-
-```
-
-  val `csw-framework`  = "$organization$" %% "csw-framework"  % Version
-
-```
-
-To start $name$ Assembly and HCD, follow below steps:
-
- - Run `sbt $name;format="space,norm"$-deploy/universal:packageBin`, this will create self contained zip in target/universal directory
- - Unzip generate zip and enter into bin directory
- - Run container cmd script or host config app script
- - Ex.  `./$name;format="word"$-host-config-app --local ../../../../$name;format="space,norm"$-deploy/src/main/resources/$name; format="Camel"$HostConfig.conf -s ./$name; format="word"$-container-cmd-app`
-
-Note: the CSW Location Service must be running before starting the components.
-See https://tmtsoftware.github.io/csw/apps/cswlocationserver.html .
