--- src/main/g8/$name__space,norm$-deploy/src/main/resources/$hcdComponentName__space,Camel$Standalone.conf
@@ -1,4 +1,0 @@
-prefix = "$subsystem$.$hcdComponentName;format="package"$"
-componentType = hcd
-behaviorFactoryClassName = "$organization$.$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$.$hcdComponentName;format="space,Camel"$BehaviorFactory"
-locationServiceUsage = RegisterOnly
