--- src/main/g8/$name__space,norm$-deploy/src/main/resources/J$hcdComponentName__space,Camel$Standalone.conf
@@ -1,4 +1,0 @@
-prefix = "$subsystem$.$hcdComponentName;format="package"$"
-componentType = hcd
-behaviorFactoryClassName = "$organization$.$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$.J$hcdComponentName;format="space,Camel"$BehaviorFactory"
-locationServiceUsage = RegisterOnly
