--- src/main/g8/$name__space,norm$-deploy/src/main/resources/J$name__space,Camel$Container.conf
@@ -1,22 +1,0 @@
-name = "$name;format="space,Camel"$Container"
-components: [
-  {
-    prefix = "$subsystem$.$assemblyComponentName;format="package"$"
-    componentType = assembly
-    behaviorFactoryClassName = "$organization$.$subsystem;format="lower"$.$assemblyComponentName;format="space,norm,word"$.J$assemblyComponentName;format="space,Camel"$BehaviorFactory"
-    locationServiceUsage = RegisterAndTrackServices
-    connections = [
-      {
-        prefix: "$subsystem$.$hcdComponentName;format="package"$"
-        componentType: hcd
-        connectionType: akka
-      }
-    ]
-  },
-  {
-    prefix = "$subsystem$.$hcdComponentName;format="package"$"
-    componentType = hcd
-    behaviorFactoryClassName = "$organization$.$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$.J$hcdComponentName;format="space,Camel"$BehaviorFactory"
-    locationServiceUsage = RegisterOnly
-  }
-]
