--- src/main/g8/$name__space,norm$-deploy/src/main/scala/$organization__packaged$/$subsystem__lower$/$name__space,norm$deploy/$name__space,Camel$HostConfigApp.scala
@@ -1,10 +1,0 @@
-package $organization$.$subsystem;format="lower"$.$name;format="space,norm,word"$deploy
-
-import csw.framework.deploy.hostconfig.HostConfig
-import csw.prefix.models.Subsystem
-
-object $name;format="space,Camel"$HostConfigApp extends App {
-
-   HostConfig.start("$name;format="normalize"$_host_config_app",Subsystem.withNameInsensitive("$subsystem$"), args)
-
-}
