+++ src/main/g8/$subsystem__lower$-$assemblyComponentName__space,norm$/src/main/resources/reference.conf
