+++ src/main/g8/$subsystem__lower$-$assemblyComponentName__space,norm$/src/test/resources/J$assemblyComponentName__space,Camel$Standalone.conf
@@ -1,0 +1,4 @@
+prefix = "$subsystem$.$assemblyComponentName;format="package"$"
+componentType = assembly
+behaviorFactoryClassName = "$subsystem;format="lower"$.$assemblyComponentName;format="space,norm,word"$.J$assemblyComponentName;format="space,Camel"$BehaviorFactory"
+locationServiceUsage = RegisterOnly
