+++ src/main/g8/$subsystem__lower$-$assemblyComponentName__space,norm$/src/test/resources/application.conf
