+++ src/main/g8/$subsystem__lower$-$hcdComponentName__space,norm$/src/main/java/$subsystem__lower$/$hcdComponentName__space,norm$/J$hcdComponentName__space,Camel$BehaviorFactory.java
@@ -1,0 +1,16 @@
+package $subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$;
+
+import akka.actor.typed.javadsl.ActorContext;
+import csw.framework.javadsl.JComponentBehaviorFactory;
+import csw.framework.javadsl.JComponentHandlers;
+import csw.framework.models.JCswContext;
+import csw.command.client.messages.TopLevelActorMessage;
+
+public class J$hcdComponentName;format="space,Camel"$BehaviorFactory extends JComponentBehaviorFactory {
+
+    @Override
+    public JComponentHandlers jHandlers(ActorContext<TopLevelActorMessage> ctx,JCswContext cswCtx) {
+        return new J$hcdComponentName;format="space,Camel"$Handlers(ctx, cswCtx);
+    }
+
+}
