+++ src/main/g8/$subsystem__lower$-$hcdComponentName__space,norm$/src/test/resources/$hcdComponentName__space,Camel$Standalone.conf
@@ -1,0 +1,4 @@
+prefix = "$subsystem$.$hcdComponentName;format="package"$"
+componentType = hcd
+behaviorFactoryClassName = "$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$.$hcdComponentName;format="space,Camel"$BehaviorFactory"
+locationServiceUsage = RegisterOnly
