+++ src/main/g8/$subsystem__lower$-$hcdComponentName__space,norm$/src/test/resources/application.conf
