+++ src/main/g8/$subsystem__lower$-$name__space,norm$-deploy/src/main/resources/$name__space,Camel$Container.conf
@@ -1,0 +1,22 @@
+name = "$name;format="space,Camel"$Container"
+components: [
+  {
+    prefix = "$subsystem$.$assemblyComponentName;format="package"$"
+    componentType = assembly
+    behaviorFactoryClassName = "$subsystem;format="lower"$.$assemblyComponentName;format="space,norm,word"$.$assemblyComponentName;format="space,Camel"$BehaviorFactory"
+    locationServiceUsage = RegisterAndTrackServices
+    connections = [
+      {
+        prefix: "$subsystem$.$hcdComponentName;format="package"$"
+        componentType: hcd
+        connectionType: akka
+      }
+    ]
+  },
+  {
+    prefix = "$subsystem$.$hcdComponentName;format="package"$"
+    componentType = hcd
+    behaviorFactoryClassName = "$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$.$hcdComponentName;format="space,Camel"$BehaviorFactory"
+    locationServiceUsage = RegisterOnly
+  }
+]
