+++ src/main/g8/$subsystem__lower$-$name__space,norm$-deploy/src/main/resources/$name__space,Camel$HostConfig.conf
