+++ src/main/g8/$subsystem__lower$-$name__space,norm$-deploy/src/main/resources/J$hcdComponentName__space,Camel$Standalone.conf
@@ -1,0 +1,4 @@
+prefix = "$subsystem$.$hcdComponentName;format="package"$"
+componentType = hcd
+behaviorFactoryClassName = "$subsystem;format="lower"$.$hcdComponentName;format="space,norm,word"$.J$hcdComponentName;format="space,Camel"$BehaviorFactory"
+locationServiceUsage = RegisterOnly
