+++ src/main/g8/$subsystem__lower$-$name__space,norm$-deploy/src/main/resources/application.conf
