+++ src/main/g8/default.properties
@@ -2,6 +2,5 @@ name=sample
 description=Template for writing TMT components (HCD's and Assembly's).  Use the name property to set the name for your system and provide default names for your components.  Words in the name property can be separated with spaces, to get correct CamelCase naming of files and classes.  Do not put spaces or dashes in component names. 
 assemblyComponentName=$name;format="package"$Assembly
 hcdComponentName=$name;format="package"$Hcd
-organization=org.tmt
 subsystem=CSW
 verbatim=coursier csw-services.sh
