+++ src/main/g8/project/Common.scala
@@ -10,7 +10,7 @@ object Common extends AutoPlugin {
   override def requires: Plugins = JvmPlugin
 
   override lazy val projectSettings: Seq[Setting[_]] = Seq(
-    organization := "$organization$",
+    organization := "com.github.tmtsoftware.$name;format="space,norm"$",
     organizationName := "TMT",
     scalaVersion := Libs.ScalaVersion,
     organizationHomepage := Some(url("http://www.tmt.org")),
